+++ src/main/java/com/motorclinic/services/interfaces/ProductService.java
@@ -7,8 +7,8 @@ import java.util.List;
 public interface ProductService {
     Product createProduct(Product product);
     Product updateProduct(Product product);
-    void deleteProduct(Integer id);
-    Product getProductById(Integer id);
+    void deleteProduct(Long id);
+    Product getProductById(Long id);
     Product getProductByName(String name);
     List<Product> getAllProducts();
 }
